feat: new folder interfaces for models
